+++ js/src/Application.js
@@ -5,6 +5,8 @@ import { BrowserRouter } from 'react-router-dom';
 import { theme, GlobalCss } from '@anu/theme';
 import { create } from 'jss';
 import rtl from 'jss-rtl';
+import { Detector } from 'react-detect-offline';
+import onNetworkChange from '@anu/utilities/onNetworkChange';
 
 // Setup JSS for RTL
 const jss = create({ plugins: [...jssPreset().plugins, rtl()] });
@@ -13,7 +15,11 @@ const Application = ({ children }) => (
   <ThemeProvider theme={theme}>
     <StylesProvider jss={jss}>
       <GlobalCss />
-      <BrowserRouter>{children}</BrowserRouter>
+      <Detector
+        polling={false}
+        onChange={onNetworkChange}
+        render={() => <BrowserRouter>{children}</BrowserRouter>}
+      />
     </StylesProvider>
   </ThemeProvider>
 );