+++ js/src/components/DownloadCourse.js
@@ -1,3 +1,4 @@
+// Legacy download course button. Consider DownloadCoursePopup.js instead.
 import React from 'react';
 import PropTypes from 'prop-types';
 import { coursePropTypes } from '@anu/utilities/transform.course';
@@ -159,6 +160,8 @@ class DownloadCourse extends React.Component {
 
       // Getting lessons and quizzes urls.
       const lessonUrls = course.content_urls;
+
+      // TODO: add lastLesson.finishButtonUrl.
 
       // Cache lessons and returns list of lesson images (from paragraphs) to cache.
       const lessonImageUrls = await this.cacheLessonsAndReturnLessonImages(lessonUrls);