+++ js/src/pages/courses/Section.js
@@ -13,7 +13,7 @@ import LinearProgress from '@material-ui/core/LinearProgress';
 import LockIcon from '@material-ui/icons/Lock';
 import TrophyIcon from '@material-ui/icons/EmojiEvents';
 import Avatar from '@material-ui/core/Avatar';
-import DownloadCourse from '@anu/components/DownloadCourse';
+import DownloadCoursePopup from '@anu/components/DownloadCoursePopup';
 import CoursesSectionEmpty from '@anu/pages/courses/SectionEmpty';
 import { coursePropTypes } from '@anu/utilities/transform.course';
 import { getPwaSettings } from '@anu/utilities/settings';
@@ -35,6 +35,9 @@ const useStyles = makeStyles((theme) => ({
     '&.locked': {
       backgroundColor: theme.palette.grey[200],
     },
+  },
+  content: {
+    paddingBottom: 0,
   },
   preTitle: {
     marginRight: theme.spacing(1),
@@ -91,9 +94,6 @@ const CoursesSection = ({ courses }) => {
     return <CoursesSectionEmpty />;
   }
 
-  const isCompleted = (course) => course.progress === '100';
-  const isUnstarted = (course) => course.progress === '0';
-
   return (
     <Grid container spacing={4}>
       {courses.map((course) => (
@@ -114,7 +114,7 @@ const CoursesSection = ({ courses }) => {
                 />
               )}
 
-              {isCompleted(course) && (
+              {course.progress_percent === 100 && (
                 <Avatar className={classes.completed}>
                   <TrophyIcon className={classes.trophyIcon} />
                 </Avatar>
@@ -123,19 +123,23 @@ const CoursesSection = ({ courses }) => {
                 <Box mb={2} display="flex" flexWrap="wrap">
                   <Typography
                     variant="overline"
-                    className={`${classes.preTitle} ${isCompleted(course) && 'completed'}`}
+                    className={`${classes.preTitle} ${
+                      course.progress_percent === 100 && 'completed'
+                    }`}
                   >
-                    {isCompleted(course)
+                    {course.progress_percent === 100
                       ? Drupal.t('Completed course', {}, { context: 'ANU LMS' })
                       : Drupal.t('Course', {}, { context: 'ANU LMS' })}
                   </Typography>
-                  {course.progress && !isCompleted(course) && !isUnstarted(course) && (
-                    <LinearProgress
-                      className={classes.progress}
-                      variant="determinate"
-                      value={Number.parseInt(course.progress, 10)}
-                    />
-                  )}
+                  {course.progress &&
+                    course.progress_percent > 0 &&
+                    course.progress_percent < 100 && (
+                      <LinearProgress
+                        className={classes.progress}
+                        variant="determinate"
+                        value={course.progress_percent}
+                      />
+                    )}
                   {course.locked && <LockIcon className={classes.lockIcon} />}
                 </Box>
 
@@ -160,7 +164,7 @@ const CoursesSection = ({ courses }) => {
 
             {course && getPwaSettings() && (
               <CardActions>
-                <DownloadCourse course={course} messagePosition="top" />
+                <DownloadCoursePopup course={course} showButton={true} />
               </CardActions>
             )}
           </Card>